refactor: split tool services
